refactor: refactor directory structure
