+++ __init__.py
@@ -1,0 +1,5 @@
+"""
+Root package marker for OrigINsight.
+This file exists to make Python treat the root directory as a package,
+allowing for absolute imports from the src directory.
+"""
