+++ src/__init__.py
@@ -1,0 +1,5 @@
+"""
+OrigINsight - A source retrieval and claim verification system.
+"""
+
+__version__ = "0.1.0"
