+++ src/core/__init__.py
@@ -1,0 +1,5 @@
+"""
+Core functionality for source retrieval and verification.
+"""
+
+from .source_retrieval import SourceRetriever, ArticleInfo
