+++ src/ui/__init__.py
@@ -1,0 +1,5 @@
+"""
+User interface implementations.
+"""
+
+from .streamlit_app import SourceRetrieverUI
